requirements.txt
ultralytics>=8.0.0
sahi>=0.11.0
streamlit>=1.30.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
